Prevent LEDs from quickly blinking when on_time == 0 or off_time == 0
In some programs a situation when, under certain conditions, LEDs should blink, while in other conditions they should be in the same state (on or off) all the time can appear. In the current code even in situation when on_time or off_time are 0, LEDs would briefly change states (first to the state with time 0 and then back on the next call to updateBlinkLed()).